+++ shared/src/commonMain/kotlin/com/perrigogames/life4/data/MessageOfTheDay.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MessageOfTheDay(
-    @SerialName("message") override val version: Int,
+    @SerialName("message") override val version: Long,
     @SerialName("image_url") val imageUrl: String,
     val body: String,
     @SerialName("white_close") val useWhiteCloseIcon: Boolean = false,
